mr_reviewer/requirements.txt
langchain==0.1.16
langchain-groq==0.1.3
requests
langchain-community==0.0.34
PyYAML>=5.3
llama-index
fuzzywuzzy
llama-index-llms-groq
python-dotenv